Fix all clippy -D warnings errors (SF-00135, SF-00137)
- unnecessary_map_or → direct comparison (orphans.rs)
- unnecessary_cast → remove as u32 (docker.rs)
- clone_on_copy → remove .clone() on Status (tui/app.rs, 3 sites)
- field_reassign_with_default → struct init syntax (config/types.rs)
- too_many_arguments → introduce StartOptions struct (start.rs)
- single_match → if-let (scripts_test.rs)

+++ src/service/docker.rs
@@ -890,7 +890,7 @@ impl ServiceManager for DockerService {
 
                 if attempt < MAX_RETRIES {
                     // Wait before retry with exponential backoff
-                    let delay_ms = 100 * 2u64.pow(attempt as u32);
+                    let delay_ms = 100 * 2u64.pow(attempt);
                     tracing::warn!(
                         "Failed to remove container {} (attempt {}/{}): {}. Retrying in {}ms...",
                         id,
